+++ libft/src/math/ft_abs.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-int ft_abs(int n)
+long ft_abs(long n)
 {
 	if (n < 0)
 		return (-n);
